src/main/java/com/programeric/HomeController.java
package com.programeric;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.servlet.ModelAndView;

@Controller
public class HomeController {
	
	@Autowired Person person;
	
	@RequestMapping("/home")
	public ModelAndView home() {
		ModelAndView mav = new ModelAndView();
		mav.addObject("message", "All systems optimal.");
		mav.setViewName("home");
		mav.addObject("person", person);
		return mav;
	}

}
